+++ alacritty/src/cli.rs
@@ -1,4 +1,5 @@
 use std::cmp::max;
+use std::os::raw::c_ulong;
 use std::path::PathBuf;
 
 #[cfg(unix)]
@@ -25,7 +26,7 @@ pub struct Options {
     #[clap(long)]
     pub ref_test: bool,
 
-    /// Defines the X11 window ID (as a decimal integer) to embed Alacritty within.
+    /// X11 window ID to embed Alacritty within (decimal or hexadecimal with "0x" prefix).
     #[clap(long)]
     pub embed: Option<String>,
 
@@ -100,7 +101,7 @@ impl Options {
         }
 
         config.window.dynamic_title &= self.window_options.window_identity.title.is_none();
-        config.window.embed = self.embed.as_ref().and_then(|embed| embed.parse().ok());
+        config.window.embed = self.embed.as_ref().and_then(|embed| parse_hex_or_decimal(embed));
         config.debug.print_events |= self.print_events;
         config.debug.log_level = max(config.debug.log_level, self.log_level());
         config.debug.ref_test |= self.ref_test;
@@ -171,6 +172,14 @@ fn parse_class(input: &str) -> Result<Class, String> {
         },
         None => Ok(Class { instance: input.into(), general: DEFAULT_NAME.into() }),
     }
+}
+
+/// Convert to hex if possible, else decimal
+fn parse_hex_or_decimal(input: &str) -> Option<c_ulong> {
+    input
+        .strip_prefix("0x")
+        .and_then(|value| c_ulong::from_str_radix(value, 16).ok())
+        .or_else(|| input.parse().ok())
 }
 
 /// Terminal specific cli options which can be passed to new windows via IPC.
@@ -391,6 +400,24 @@ mod tests {
     fn parse_invalid_class() {
         let class = parse_class("one,two,three");
         assert!(class.is_err());
+    }
+
+    #[test]
+    fn valid_decimal() {
+        let value = parse_hex_or_decimal("10485773");
+        assert_eq!(value, Some(10485773));
+    }
+
+    #[test]
+    fn valid_hex_to_decimal() {
+        let value = parse_hex_or_decimal("0xa0000d");
+        assert_eq!(value, Some(10485773));
+    }
+
+    #[test]
+    fn invalid_hex_to_decimal() {
+        let value = parse_hex_or_decimal("0xa0xx0d");
+        assert_eq!(value, None);
     }
 
     #[cfg(target_os = "linux")]